Switch back to cloud functions for GTFS-RT archiving
