+++ iac/cal-itp-data-infra-staging/gtfs-rt-archiver/us/variables.tf
@@ -3,21 +3,6 @@ locals {
   archive_path = "${path.module}/build/source.zip"
 }
 
-data "external" "git" {
-  program = [
-    "git",
-    "log",
-    "--pretty=format:{ \"sha\": \"%H\" }",
-    "-1",
-    "HEAD"
-  ]
-}
-
-variable "git_sha" {
-  type    = string
-  default = null
-}
-
 data "terraform_remote_state" "iam" {
   backend = "gcs"
