--- services/gtfs-rt-archiver/gtfs_rt_archiver/logger.py
@@ -1,26 +1,0 @@
-import logging
-import sys
-
-import google.cloud.logging
-
-
-class LogLevelFilter(logging.Filter):
-    def filter(self, rec):
-        return rec.levelno in (logging.DEBUG, logging.INFO)
-
-
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.DEBUG)
-
-stdout_handler = logging.StreamHandler(sys.stdout)
-stdout_handler.setLevel(logging.DEBUG)
-stdout_handler.addFilter(LogLevelFilter())
-logger.addHandler(stdout_handler)
-
-stderr_handler = logging.StreamHandler()
-stderr_handler.setLevel(logging.WARNING)
-
-logger.addHandler(stderr_handler)
-
-client = google.cloud.logging.Client()
-client.setup_logging()
