+++ services/gtfs-rt-archiver/requirements.txt
@@ -1,0 +1,4 @@
+functions-framework==3.*
+google-cloud-secret-manager==2.*
+google-cloud-storage==3.*
+google-cloud-logging==3.*
