+++ services/gtfs-rt-archiver/tests/test_service.py
@@ -4,12 +4,11 @@ from datetime import datetime
 
 import pytest
 from google.cloud.storage import Blob, Bucket, Client
-from gtfs_rt_archiver.archiver import Archiver
 from gtfs_rt_archiver.configuration import Configuration
-from gtfs_rt_archiver.downloader import Downloader
+from gtfs_rt_archiver.service import Service
 
 
-class TestArchiver:
+class TestService:
     @pytest.fixture
     def publish_time(self) -> datetime:
         return datetime.fromisoformat("2026-04-01T00:01:23.45+00:00")
@@ -19,9 +18,8 @@ class TestArchiver:
         return "http://avl.yctd.org/RealTime/GTFS_ServiceAlerts.pb"
 
     @pytest.fixture
-    def data(self, publish_time: datetime, url: str) -> dict:
+    def data(self, url: str) -> dict:
         return {
-            "publish_time": publish_time,
             "auth_headers": {},
             "auth_query_params": {},
             "extracted_at": "2026-04-01T00:00:00+00:00",
@@ -32,16 +30,12 @@ class TestArchiver:
         }
 
     @pytest.fixture
-    def configuration(self, data: dict) -> Configuration:
-        return Configuration.resolve(**data)
+    def configuration(self, data: dict, publish_time: datetime) -> Configuration:
+        return Configuration.resolve(publish_time=publish_time.isoformat(), **data)
 
     @pytest.fixture
-    def downloader(self, configuration: Configuration) -> Downloader:
-        return Downloader(configuration=configuration)
-
-    @pytest.fixture
-    def archiver(self, configuration: Configuration) -> Archiver:
-        return Archiver(configuration=configuration)
+    def service(self, data: dict, publish_time: datetime) -> Service:
+        return Service(data=json.dumps(data), publish_time=publish_time, message_id="1")
 
     @pytest.fixture
     def client(self) -> Client:
@@ -69,10 +63,8 @@ class TestArchiver:
         return bucket.get_blob(blob_name=destination_path)
 
     @pytest.mark.vcr
-    def test_archiver_stores_feed(
-        self, downloader: Downloader, archiver: Archiver, blob: Blob
-    ) -> None:
-        archiver.save(result=downloader.get())
+    def test_service_downloads_blob(self, service: Service, blob: Blob) -> None:
+        service.run()
         assert json.loads(blob.metadata["PARTITIONED_ARTIFACT_METADATA"]) == {
             "filename": "feed",
             "ts": "2026-04-01T00:01:20+00:00",
@@ -87,13 +79,13 @@ class TestArchiver:
             },
             "response_code": 200,
             "response_headers": {
-                "Last-Modified": "Tue, 07 Apr 2026 22:36:11 GMT",
-                "Date": "Tue, 07 Apr 2026 22:36:16 GMT",
+                "Last-Modified": "Tue, 07 Apr 2026 22:36:26 GMT",
+                "Date": "Tue, 07 Apr 2026 22:36:27 GMT",
                 "Server": "Microsoft-IIS/10.0",
                 "Content-Type": "application/x-protobuf",
                 "X-Powered-By": "ASP.NET",
                 "Accept-Ranges": "bytes",
                 "Content-Length": "2960",
-                "ETag": '"422da1efdec6dc1:0"',
+                "ETag": 'W/"556a91f8dec6dc1:0"',
             },
         }
